+++ api/src/matches/matches.controller.ts
@@ -1,21 +1,16 @@
-import { Body, Controller, Get, Param, Patch, Post } from '@nestjs/common';
+import { Body, Controller, Get, Param, Patch, Post, UseGuards } from '@nestjs/common';
+import { JwtAuthGuard } from '../auth/jwt.auth.guard';
 import { CreateMatchDto } from './dto/create-match.dto';
 import { UpdateMatchDto } from './dto/update-match.dto';
 import { MatchesService } from './matches.service';
 
 @Controller('matches')
 export class MatchesController {
-  constructor(private readonly matchesService: MatchesService) { }
-
   /**
-   * Creates a past or future match.
-   * @param createMatchDto 
-   * @returns The created match.
+   * Constructor.
+   * @param matchesService The matches service. 
    */
-  @Post()
-  create(@Body() createMatchDto: CreateMatchDto) {
-    return this.matchesService.create(createMatchDto);
-  }
+  constructor(private readonly matchesService: MatchesService) { }
 
   /**
    * Finds all the matches.
@@ -25,7 +20,24 @@ export class MatchesController {
   findAll() {
     return this.matchesService.findAll();
   }
+
+  /**
+ * Creates a past or future match.
+ * @param createMatchDto The match creation validator.
+ * @returns The created match.
+ */
+  @Post()
+  @UseGuards(JwtAuthGuard)
+  create(@Body() createMatchDto: CreateMatchDto) {
+    return this.matchesService.create(createMatchDto);
+  }
+
+  /**
+   * Updates a match.
+   * @param updateMatchDto The match update validator.
+   */
   @Patch(':id')
+  @UseGuards(JwtAuthGuard)
   update(@Param('id') id: string, @Body() updateMatchDto: UpdateMatchDto) {
     return this.matchesService.update(id, updateMatchDto);
   }