+++ app/src/app/app.routes.ts
@@ -1,6 +1,7 @@
 import { Routes } from '@angular/router';
-import { ConsultationComponent } from './pages/consultation/consultation/consultation.component';
-import { LoginComponent } from './pages/login/login/login.component';
+import { ConsultationComponent } from './pages/consultation/consultation.component';
+import { LoginComponent } from './pages/login/login.component';
+import { authGuard } from './shared/auth/auth.guard';
 import { APP_ROUTES } from './shared/constants/routes.constants';
 
 /**
@@ -12,7 +13,8 @@ export const routes: Routes = [
     { path: APP_ROUTES.consultation, component: ConsultationComponent },
     {
         path: 'admin',
-        loadChildren: () => import('./pages/admin/admin.routes')
+        loadChildren: () => import('./pages/admin/admin.routes'),
+        canActivate: [authGuard]
     },
     { path: '**', redirectTo: APP_ROUTES.consultation }
 ];