+++ app/src/app/pages/admin/teams/create-team.component.ts
@@ -1,5 +1,8 @@
 import { Component } from '@angular/core';
 
+/**
+ * The teams management page.
+ */
 @Component({
   selector: 'app-create-team',
   standalone: true,