+++ app/src/app/pages/consultation/consultation.component.ts
@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
+import { ButtonComponent } from "../../shared/components/button/button.component";
 
 @Component({
   selector: 'app-consultation',
   standalone: true,
-  imports: [],
+  imports: [ButtonComponent],
   templateUrl: './consultation.component.html',
   styleUrl: './consultation.component.scss'
 })