+++ app/CHANGELOG.md
@@ -8,3 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.1.0
 
 Initial release.
+
+### Added
+
+- Pages and routing.
+- Simple authentication IHM, guard.